move defs to new source file
